+++ data/catData.ts
@@ -1,4 +1,4 @@
-import { Cat } from 'store/cat';
+import { Cat } from '@/src/types/cat.type';
 
 export const catLibrary: Cat[] = [
   { id: '1', name: 'Garfield', color: 'ginger' },
