+++ src/store/state-management/redux/posts/selectors.ts
@@ -2,7 +2,7 @@
 
 import { EntityAdapter } from '@reduxjs/toolkit';
 
-import { Post } from 'store/posts';
+import { Post } from 'store/post.type';
 
 import { useAppSelector } from '../hook';
 import { RootState } from '../store';
