+++ src/store/state-management/redux/posts/slice.ts
@@ -16,7 +16,7 @@ import {
   createEntityAdapter,
   createSlice,
 } from '@reduxjs/toolkit';
-import { Post } from 'store/posts';
+import { Post } from 'store/post.type';
 
 export type InitialStateType = {
   status: 'idle' | 'loading' | 'succeeded' | 'failed';
