--- src/store/transport-layer.ts
@@ -1,3 +1,0 @@
-export type PostTransportLayer = {
-  fetchPosts: () => Promise<Response>;
-};
