+++ src/types/cat.type.ts
