+++ src/types/music.type.ts
