+++ src/types/product.type.ts
